requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
scipy>=1.11.0
jupyter>=1.0.0
notebook>=7.0.0
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0